[88] Created a genric list page component and applied it to pages that already existed
