+++ src/Pages/ClientListScreen/index.js
@@ -1,12 +1,10 @@
 import React, { useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
-import { FaSistrix } from 'react-icons/fa';
 import axios from 'axios';
-import SearchInput from '../../Components/SearchInput';
+import { Link } from 'react-router-dom';
 import ClientProfileData from '../../Components/ClientProfileData';
+import GenericListScreen from '../../Components/GenericListScreen';
 import {
-  Main, Container, Header, Title, Search, TableHeader, P, Bar,
-  DataContainer, ButtonDiv, TableTitle, ContentBox,
+  TableHeader, P, Bar, TableTitle,
 } from './style';
 
 const ClientListScreen = () => {
@@ -51,63 +49,47 @@ const ClientListScreen = () => {
   };
 
   return (
-    <Main>
-      <Container>
-        <Header>
-          <Title>Clientes</Title>
-          <Search>
-            <SearchInput
-              type="text"
-              icon={<FaSistrix />}
-              value={word}
-              setWord={(value) => setWord(value)}
-            />
-          </Search>
-          <ButtonDiv>
-            <Link
-              to="/cliente"
-              style={{
-                color: 'white',
-                textDecorationLine: 'none',
-                fontSize: '1.5vw',
-              }}
-            >
-              Novo cliente
-            </Link>
-          </ButtonDiv>
-        </Header>
+    <GenericListScreen
+      ButtonTitle="Novo Cliente"
+      ButtonFunction={() => {
+        <Link
+          to="/cliente"
+          style={{
+            color: 'white',
+            textDecorationLine: 'none',
+            fontSize: '1.5vw',
+          }}
+        />;
+      }}
+      PageTitle="Clientes"
+      SearchWord={word}
+      setWord={setWord}
+      ListType={listClients()}
+    >
+      <TableHeader>
+        <TableTitle width={25}>
+          <P>Nome</P>
+        </TableTitle>
+        <Bar />
+        <TableTitle width={25}>
+          <P>Email</P>
+        </TableTitle>
+        <Bar />
 
-        <ContentBox>
-          <TableHeader>
-            <TableTitle width={30}>
-              <P>Nome</P>
-            </TableTitle>
-            <Bar />
-            <TableTitle width={20}>
-              <P>Email</P>
-            </TableTitle>
-            <Bar />
+        <TableTitle width={15}>
+          <P>CPF</P>
+        </TableTitle>
+        <Bar />
 
-            <TableTitle width={15}>
-              <P>CPF</P>
-            </TableTitle>
-            <Bar />
-
-            <TableTitle width={15}>
-              <P>Telefone</P>
-            </TableTitle>
-            <Bar />
-            <TableTitle width={19}>
-              <P>Ult. Atualização</P>
-            </TableTitle>
-          </TableHeader>
-
-          <DataContainer>
-            {listClients()}
-          </DataContainer>
-        </ContentBox>
-      </Container>
-    </Main>
+        <TableTitle width={15}>
+          <P>Telefone</P>
+        </TableTitle>
+        <Bar />
+        <TableTitle width={19}>
+          <P>Ult. Atualização</P>
+        </TableTitle>
+      </TableHeader>
+    </GenericListScreen>
   );
 };
